update card and toast examples
