+++ packages/sit-onyx/src/components/OnyxCard/examples/DetailsCardExample.vue
