+++ packages/sit-onyx/src/components/OnyxCard/examples/HeadlineCardExample.vue
