+++ packages/sit-onyx/src/components/OnyxCard/examples/ImageCardExample.vue
