+++ packages/sit-onyx/src/components/OnyxCard/examples/KPICardExample.vue
